Moved validation classes to the Validators folder. Updated the validation classes so that console messages clearly indicate which validation test we are running.
